+++ src/lib/imem.h
@@ -6,15 +6,15 @@ extern Pool default_pool;
 /* For easy allocation of memory from default memory pool. */
 #define i_new(type, count) \
         ((type *) i_malloc((unsigned) sizeof(type) * (count)))
-void *i_malloc(unsigned int size);
+void *i_malloc(size_t size);
 void i_free(void *mem);
-void *i_realloc(void *mem, unsigned int size);
+void *i_realloc(void *mem, size_t size);
 
 /* string functions */
 char *i_strdup(const char *str);
 char *i_strdup_empty(const char *str); /* like i_strdup(), but if str == "", return NULL */
 char *i_strdup_until(const char *str, const char *end); /* *end isn't included */
-char *i_strndup(const char *str, unsigned int max_chars);
+char *i_strndup(const char *str, size_t max_chars);
 char *i_strdup_printf(const char *format, ...) __attr_format__(1, 2);
 char *i_strdup_vprintf(const char *format, va_list args);
 void i_strdup_replace(char **dest, const char *str);
