+++ src/lib/mempool.c
@@ -26,9 +26,9 @@
 #include "lib.h"
 #include "mempool.h"
 
-Pool pool_alloconly_create(const char *name, unsigned int size);
+Pool pool_alloconly_create(const char *name, size_t size);
 
-Pool pool_create(const char *name, unsigned int size, int allocfree)
+Pool pool_create(const char *name, size_t size, int allocfree)
 {
 	if (allocfree)
 		return system_pool;
